Fixed the expander
